Separated node class from osm_network_creator
